pull Lambda configuration out of artifact_builder template module
