+++ cli/src/accretion_cli/__init__.py
@@ -1,2 +1,6 @@
 """Accretion CloudFormation template artifact_builder."""
+from troposphere import Tags
+
 __version__ = "0.0.1"
+
+DEFAULT_TAGS = Tags(Accretion="ArtifactBuilder")
