+++ cli/src/accretion_cli/_templates/artifact_builder.py
@@ -1,63 +1,13 @@
 """Build the CloudFormation template for the Accretion artifact builder."""
 from functools import partial
-from typing import Callable, Dict, Iterable
+from typing import Callable
 
-from troposphere import Template, Tags, Parameter, iam, awslambda
-from awacs import aws as AWS
+from troposphere import Template, Parameter, awslambda
 
-from accretion_cli._iam import lambda_role, s3_put_object_statement
+from accretion_cli import DEFAULT_TAGS
+from accretion_cli._iam import s3_put_object_statement
+from accretion_cli._lambda import lambda_function, add_lambda_core
 from accretion_cli._stepfunctions import add_artifact_builder
-
-
-_DEFAULT_TAGS = Tags(Accretion="ArtifactBuilder")
-
-
-def _lambda_environment(bucket_name: Parameter) -> awslambda.Environment:
-    return awslambda.Environment(Variables=dict(S3_BUCKET=bucket_name.ref()))
-
-
-def _lambda_function(
-    *,
-    bucket_name: Parameter,
-    workers_key: Parameter,
-    boto3_layer: Parameter,
-    name: str,
-    role: iam.Role,
-    runtime: str,
-    module: str,
-    memory_size: int,
-    timeout: int,
-) -> awslambda.Function:
-    return awslambda.Function(
-        name,
-        Role=role.get_att("Arn"),
-        Code=awslambda.Code(S3Bucket=bucket_name.ref(), S3Key=workers_key.ref()),
-        Handler=f"accretion_workers.layer_builder.{module}.lambda_handler",
-        Environment=_lambda_environment(bucket_name),
-        Runtime=runtime,
-        Layers=[boto3_layer.ref()],
-        MemorySize=memory_size,
-        Timeout=timeout,
-        Tags=_DEFAULT_TAGS,
-    )
-
-
-def _add_lambda_core(
-    *,
-    builder: Template,
-    lambda_builder: Callable,
-    base_name: str,
-    statements: Iterable[AWS.Statement],
-    lambda_args: Dict,
-) -> awslambda.Function:
-    role = lambda_role(f"{base_name}Role", *statements)
-
-    function = lambda_builder(name=f"{base_name}Function", role=role, **lambda_args)
-
-    builder.add_resource(role)
-    builder.add_resource(function)
-
-    return function
 
 
 def _add_parse_requirements_resources(lambda_adder: Callable) -> awslambda.Function:
@@ -97,9 +47,9 @@ def build() -> Template:
     )
 
     _lambda_builder = partial(
-        _lambda_function, bucket_name=bucket_name, workers_key=workers_key, boto3_layer=boto3_layer
+        lambda_function, bucket_name=bucket_name, workers_key=workers_key, boto3_layer=boto3_layer
     )
-    lambda_adder = partial(_add_lambda_core, builder=builder, lambda_builder=_lambda_builder)
+    lambda_adder = partial(add_lambda_core, builder=builder, lambda_builder=_lambda_builder)
 
     parse_requirements = _add_parse_requirements_resources(lambda_adder)
 
@@ -112,7 +62,7 @@ def build() -> Template:
         parse_requirements=parse_requirements,
         python36_builder=python36_builder,
         python37_builder=python37_builder,
-        tags=_DEFAULT_TAGS,
+        tags=DEFAULT_TAGS,
     )
 
     return builder
